+++ projects/modals/src/modules/confirm/confirm.service.ts
@@ -38,7 +38,9 @@ import {
  *  The `SkyConfirmService` provider launches the confirmation dialog.
  * Within the service, you can specify the dialog's message and customize the button text.
  */
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class SkyConfirmService {
   constructor(
     private modalService: SkyModalService