+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 5.0.0-beta.1 (2021-07-26)
+
+- Fixed the `SkyModalService` to work within lazy-loaded modules. [#160](https://github.com/blackbaud/skyux-modals/pull/160)
+
+# 5.0.0-beta.0 (2021-07-06)
+
+- Initial beta release.
+- Added support for `5.0.0-beta.*` versions of SKY UX component libraries. [#155](https://github.com/blackbaud/skyux-modals/pull/155)
+
+# 4.5.4 (2021-06-14)
+
+- Fixed a bug where `SkyThemeService` was erroneously provided by `SkyConfirmModule` and `SkyModalModule`. [#153](https://github.com/blackbaud/skyux-modals/pull/153)
+
 # 4.5.3 (2021-05-04)
 
 - Deprecated `SkyConfirmType.YesCancel` and `SkyConfirmType.YesNoCancel`. Use the `SkyConfirmType.Custom` type to follow the guidance that labels should clearly indicate the actions that occur when users select buttons. [#147](https://github.com/blackbaud/skyux-modals/pull/147)